XCframework changes
1- reduced size of the package
2- added new production functionality
