+++ RiskifiedBeacon.xcframework/ios-arm64/RiskifiedBeacon.framework/Headers/RiskifiedBeacon-Swift.h
@@ -1,4 +1,4 @@
-// Generated by Apple Swift version 5.6 (swiftlang-5.6.0.323.62 clang-1316.0.20.8)
+// Generated by Apple Swift version 5.6.1 (swiftlang-5.6.0.323.66 clang-1316.0.20.12)
 #ifndef RISKIFIEDBEACON_SWIFT_H
 #define RISKIFIEDBEACON_SWIFT_H
 #pragma clang diagnostic push
@@ -239,6 +239,34 @@ SWIFT_CLASS("_TtC15RiskifiedBeacon10DecoLogger")
 SWIFT_CLASS("_TtC15RiskifiedBeacon7DecoSdk")
 @interface DecoSdk : NSObject
 - (void)offerDecoWithConfig:(DecoConfig * _Nonnull)config;
+- (nonnull instancetype)init OBJC_DESIGNATED_INITIALIZER;
+@end
+
+typedef SWIFT_ENUM(NSInteger, Env, open) {
+  EnvProduction = 0,
+  EnvSandbox = 1,
+};
+
+
+SWIFT_CLASS("_TtC15RiskifiedBeacon9MfaConfig")
+@interface MfaConfig : NSObject
+- (nonnull instancetype)init SWIFT_UNAVAILABLE;
++ (nonnull instancetype)new SWIFT_UNAVAILABLE_MSG("-init is unavailable");
+@end
+
+
+SWIFT_CLASS("_TtC15RiskifiedBeacon9MfaLogger")
+@interface MfaLogger : NSObject
+- (nonnull instancetype)initWithIsDebug:(BOOL)isDebug OBJC_DESIGNATED_INITIALIZER;
+- (void)infoWithMessage:(NSString * _Nonnull)message;
+- (nonnull instancetype)init SWIFT_UNAVAILABLE;
++ (nonnull instancetype)new SWIFT_UNAVAILABLE_MSG("-init is unavailable");
+@end
+
+
+SWIFT_CLASS("_TtC15RiskifiedBeacon6MfaSDK")
+@interface MfaSDK : NSObject
+- (void)offerMFAWithConfig:(MfaConfig * _Nonnull)config;
 - (nonnull instancetype)init OBJC_DESIGNATED_INITIALIZER;
 @end
