+++ RiskifiedBeacon.xcframework/ios-arm64_x86_64-simulator/RiskifiedBeacon.framework/_CodeSignature/CodeResources
@@ -1,0 +1,201 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>files</key>
+	<dict>
+		<key>Headers/RiskifiedBeacon-Swift.h</key>
+		<data>
+		/jdBjKHd1/TDY0fir7OVTWkSEr0=
+		</data>
+		<key>Headers/RiskifiedBeacon.h</key>
+		<data>
+		HZ12Cu9XgMQwefIWim6GsipM2K0=
+		</data>
+		<key>Info.plist</key>
+		<data>
+		DB+MBnZCtGknlTIAennYZyBBWBk=
+		</data>
+		<key>Modules/RiskifiedBeacon.swiftmodule/arm64-apple-ios-simulator.swiftdoc</key>
+		<data>
+		9hbrQrREjtuI648qGNfSNlOALso=
+		</data>
+		<key>Modules/RiskifiedBeacon.swiftmodule/arm64-apple-ios-simulator.swiftinterface</key>
+		<data>
+		KFFxhkB6p5Eu0qfSt4MkQ7fAFq4=
+		</data>
+		<key>Modules/RiskifiedBeacon.swiftmodule/arm64-apple-ios-simulator.swiftmodule</key>
+		<data>
+		bVBYealYBWO6psuaxoGmbcZ+pms=
+		</data>
+		<key>Modules/RiskifiedBeacon.swiftmodule/x86_64-apple-ios-simulator.swiftdoc</key>
+		<data>
+		ZMjBjI57ls4+yjbMIVCjXlrCmTw=
+		</data>
+		<key>Modules/RiskifiedBeacon.swiftmodule/x86_64-apple-ios-simulator.swiftinterface</key>
+		<data>
+		3ZNkrWlQDYPdodm5JyUi+UooWg0=
+		</data>
+		<key>Modules/RiskifiedBeacon.swiftmodule/x86_64-apple-ios-simulator.swiftmodule</key>
+		<data>
+		kLkmGpSPZzgrlZCdz+/ztLeElW4=
+		</data>
+		<key>Modules/module.modulemap</key>
+		<data>
+		fVtd/C3/EOJjtbmJwMVFEWswcc4=
+		</data>
+	</dict>
+	<key>files2</key>
+	<dict>
+		<key>Headers/RiskifiedBeacon-Swift.h</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			b3Jdbv34g2ihgIyqkwat0vaO2tam6N1/lHadklWBTXk=
+			</data>
+		</dict>
+		<key>Headers/RiskifiedBeacon.h</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			aEYmY4tZ4a0S6cjBugvAeRbprrqMZgAP0gY+GPE2Jr8=
+			</data>
+		</dict>
+		<key>Modules/RiskifiedBeacon.swiftmodule/arm64-apple-ios-simulator.swiftdoc</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			n68zlSr32jc6oeAEiidqsDraefWN+gbJg/gz4dursuk=
+			</data>
+		</dict>
+		<key>Modules/RiskifiedBeacon.swiftmodule/arm64-apple-ios-simulator.swiftinterface</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			QBxXNkRidVyVTQWGzJKlTjF/sSNbqagy06EqaTd5XW8=
+			</data>
+		</dict>
+		<key>Modules/RiskifiedBeacon.swiftmodule/arm64-apple-ios-simulator.swiftmodule</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			X0ECrx6ol2pkfIkky+sXimEVgAq6LCRbVyVkwZ+lwoo=
+			</data>
+		</dict>
+		<key>Modules/RiskifiedBeacon.swiftmodule/x86_64-apple-ios-simulator.swiftdoc</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			yfr56ZPw0YoDl3kAc7OWuAyub/gWVBjyPMEhmNAjwXI=
+			</data>
+		</dict>
+		<key>Modules/RiskifiedBeacon.swiftmodule/x86_64-apple-ios-simulator.swiftinterface</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			7OpgVbaTqzmEZ4AFQr8cghTR1rLki89weaj4UfDiX10=
+			</data>
+		</dict>
+		<key>Modules/RiskifiedBeacon.swiftmodule/x86_64-apple-ios-simulator.swiftmodule</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			nYHyVtfJMuVGZblu4zO1mHDse+6wUk/UERSLWTKEgDM=
+			</data>
+		</dict>
+		<key>Modules/module.modulemap</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			4F50ji8Ihhq3tENGccorV5wTEVbXqAmYMy7gDRSZCeU=
+			</data>
+		</dict>
+	</dict>
+	<key>rules</key>
+	<dict>
+		<key>^.*</key>
+		<true/>
+		<key>^.*\.lproj/</key>
+		<dict>
+			<key>optional</key>
+			<true/>
+			<key>weight</key>
+			<real>1000</real>
+		</dict>
+		<key>^.*\.lproj/locversion.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>1100</real>
+		</dict>
+		<key>^Base\.lproj/</key>
+		<dict>
+			<key>weight</key>
+			<real>1010</real>
+		</dict>
+		<key>^version.plist$</key>
+		<true/>
+	</dict>
+	<key>rules2</key>
+	<dict>
+		<key>.*\.dSYM($|/)</key>
+		<dict>
+			<key>weight</key>
+			<real>11</real>
+		</dict>
+		<key>^(.*/)?\.DS_Store$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>2000</real>
+		</dict>
+		<key>^.*</key>
+		<true/>
+		<key>^.*\.lproj/</key>
+		<dict>
+			<key>optional</key>
+			<true/>
+			<key>weight</key>
+			<real>1000</real>
+		</dict>
+		<key>^.*\.lproj/locversion.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>1100</real>
+		</dict>
+		<key>^Base\.lproj/</key>
+		<dict>
+			<key>weight</key>
+			<real>1010</real>
+		</dict>
+		<key>^Info\.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^PkgInfo$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^embedded\.provisionprofile$</key>
+		<dict>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^version\.plist$</key>
+		<dict>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+	</dict>
+</dict>
+</plist>
